+++ tests/unit/MockExpectedCall.test.cpp
@@ -793,9 +793,9 @@ TEST(ExpectedCall, hasUnmodifiedOutputParameter)
 
 TEST(ExpectedCall, hasNoOutputParameter)
 {
-  call->with_int_parameter("foo", static_cast<int>(1));
+  call->with_int_parameter("foo", 1);
   mu::tiny::mock::NamedValue foo("foo");
-  foo.set_value(static_cast<int>(1));
+  foo.set_value(1);
   CHECK(!call->has_output_parameter(foo));
 }
