+++ tests/unit/MockIgnoredExpectedCall.test.cpp
@@ -14,17 +14,13 @@ TEST(IgnoredExpectedCall, worksAsItShould)
   ignored.with_call_order(1, 1);
   ignored.on_object(nullptr);
   ignored.with_bool_parameter("umm", true);
-  ignored.with_int_parameter("bla", static_cast<int>(1));
-  ignored.with_unsigned_int_parameter("foo", static_cast<unsigned int>(1));
-  ignored.with_long_int_parameter("hey", static_cast<long int>(1));
-  ignored.with_unsigned_long_int_parameter(
-      "bah", static_cast<unsigned long int>(1)
-  );
-  ignored.with_long_long_int_parameter("yo", static_cast<long long int>(1));
-  ignored.with_unsigned_long_long_int_parameter(
-      "grr", static_cast<unsigned long long int>(1)
-  );
-  ignored.with_double_parameter("hah", static_cast<double>(1.1f));
+  ignored.with_int_parameter("bla", 1);
+  ignored.with_unsigned_int_parameter("foo", 1U);
+  ignored.with_long_int_parameter("hey", 1L);
+  ignored.with_unsigned_long_int_parameter("bah", 1UL);
+  ignored.with_long_long_int_parameter("yo", 1LL);
+  ignored.with_unsigned_long_long_int_parameter("grr", 1ULL);
+  ignored.with_double_parameter("hah", 1.1);
   ignored.with_double_parameter("gah", 2.1, 0.3);
   ignored.with_string_parameter("goo", "hello");
   ignored.with_pointer_parameter("pie", static_cast<void*>(nullptr));
@@ -49,13 +45,13 @@ TEST(IgnoredExpectedCall, worksAsItShould)
   ignored.with_unmodified_output_parameter("unmod");
   ignored.ignore_other_parameters();
   ignored.and_return_value(true);
-  ignored.and_return_value(static_cast<double>(1.0f));
-  ignored.and_return_value(static_cast<unsigned int>(1));
-  ignored.and_return_value(static_cast<int>(1));
-  ignored.and_return_value(static_cast<unsigned long int>(1));
-  ignored.and_return_value(static_cast<long int>(1));
-  ignored.and_return_value(static_cast<unsigned long long int>(1));
-  ignored.and_return_value(static_cast<long long int>(1));
+  ignored.and_return_value(1.0);
+  ignored.and_return_value(1U);
+  ignored.and_return_value(1);
+  ignored.and_return_value(1UL);
+  ignored.and_return_value(1L);
+  ignored.and_return_value(1ULL);
+  ignored.and_return_value(1LL);
   ignored.and_return_value("boo");
   ignored.and_return_value(static_cast<void*>(nullptr));
   ignored.and_return_value(static_cast<const void*>(nullptr));
